egov_delivery/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-admin==2.0.2
django-cors-headers==3.14.0
django-excel-response2==3.0.5
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
drf-yasg==1.21.5
environ==1.0
excel-base==1.0.4
idna==3.4
inflection==0.5.1
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.9
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
screen==1.0.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
TimeConvert==3.0.13
tomli==2.0.1
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
xlwt==1.3.0
